ReducedLung use dummy geometry element in discret
