+++ src/global_legacy_module/4C_global_legacy_module.cpp
@@ -112,7 +112,6 @@
 #include "4C_porofluid_pressure_based_elast_scatra_function.hpp"
 #include "4C_porofluid_pressure_based_ele.hpp"
 #include "4C_red_airways_elementbase.hpp"
-#include "4C_reduced_lung_discretization_helpers.hpp"
 #include "4C_rigidsphere.hpp"
 #include "4C_scatra_ele.hpp"
 #include "4C_scatra_functions.hpp"
@@ -210,7 +209,6 @@ namespace
       << Discret::Elements::RedAirwayType::instance().name() << " "
       << Discret::Elements::RedAcinusType::instance().name() << " "
       << Discret::Elements::RedInterAcinarDepType::instance().name() << " "
-      << Discret::Elements::ReducedLungLineType::instance().name() << " "
       << Discret::Elements::ConstraintElement2Type::instance().name() << " "
       << Discret::Elements::ConstraintElement3Type::instance().name() << " "
       << Discret::Elements::LubricationType::instance().name() << " "
